RIA-5804 using feature flag (#586)
