+++ docker-compose.yml
@@ -31,6 +31,7 @@ services:
       IA_IDAM_REDIRECT_URI: http://idam-api
       IDAM_URL: http://idam-api:8080
       DM_URL: http://dm-store:8080
+      CASE_DOCUMENT_AM_URL: http://ccd-case-document-am-api:4455
       CCD_GW_URL: http://ccd-api-gateway-web:3453
       EM_BUNDLER_URL: http://rpa-em-ccd-orchestrator:8060
       DOCMOSIS_ENDPOINT:
