+++ src/functionalTest/java/uk/gov/hmcts/reform/iacasedocumentsapi/util/SystemDocumentManagementUploader.java
@@ -1,85 +1,27 @@
 package uk.gov.hmcts.reform.iacasedocumentsapi.util;
 
-import com.google.common.io.ByteStreams;
-import java.io.IOException;
-import java.util.Collections;
+import lombok.RequiredArgsConstructor;
 import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
-import uk.gov.hmcts.reform.document.DocumentUploadClientApi;
-import uk.gov.hmcts.reform.document.domain.UploadResponse;
-import uk.gov.hmcts.reform.document.utils.InMemoryMultipartFile;
 import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.ccd.field.Document;
+import uk.gov.hmcts.reform.iacasedocumentsapi.domain.service.FeatureToggler;
 
 @Service
+@RequiredArgsConstructor
 public class SystemDocumentManagementUploader {
 
-    private final DocumentUploadClientApi documentUploadClientApi;
-    private final AuthorizationHeadersProvider authorizationHeadersProvider;
+    private final FeatureToggler featureToggler;
 
-    public SystemDocumentManagementUploader(
-        DocumentUploadClientApi documentUploadClientApi,
-        AuthorizationHeadersProvider authorizationHeadersProvider
-    ) {
-        this.documentUploadClientApi = documentUploadClientApi;
-        this.authorizationHeadersProvider = authorizationHeadersProvider;
-    }
+    private final CdamSystemDocumentManagementUploader cdamSystemDocumentManagementUploader;
 
-    public Document upload(
-        Resource resource,
-        String contentType
-    ) {
-        final String serviceAuthorizationToken =
-            authorizationHeadersProvider
-                .getLegalRepresentativeAuthorization()
-                .getValue("ServiceAuthorization");
+    private final DmSystemDocumentManagementUploader dmSystemDocumentManagementUploader;
 
-        final String accessToken =
-            authorizationHeadersProvider
-                .getLegalRepresentativeAuthorization()
-                .getValue("Authorization");
-
-        final String userId = "1";
-
-        try {
-
-            MultipartFile file = new InMemoryMultipartFile(
-                resource.getFilename(),
-                resource.getFilename(),
-                contentType,
-                ByteStreams.toByteArray(resource.getInputStream())
-            );
-
-            UploadResponse uploadResponse =
-                documentUploadClientApi
-                    .upload(
-                        accessToken,
-                        serviceAuthorizationToken,
-                        userId,
-                        Collections.singletonList(file)
-                    );
-
-            uk.gov.hmcts.reform.document.domain.Document uploadedDocument =
-                uploadResponse
-                    .getEmbedded()
-                    .getDocuments()
-                    .get(0);
-
-            return new Document(
-                uploadedDocument
-                    .links
-                    .self
-                    .href,
-                uploadedDocument
-                    .links
-                    .binary
-                    .href,
-                uploadedDocument
-                    .originalDocumentName
-            );
-
-        } catch (IOException e) {
-            throw new IllegalStateException(e);
+    public Document upload(Resource resource, String contentType) {
+        if (featureToggler.getValue("use-ccd-document-am", false)) {
+            return cdamSystemDocumentManagementUploader.upload(resource, contentType);
+        } else {
+            return dmSystemDocumentManagementUploader.upload(resource, contentType);
         }
+
     }
 }
