+++ src/integrationTest/java/uk/gov/hmcts/reform/iacasedocumentsapi/component/testutils/TestClassOrderer.java
@@ -1,0 +1,17 @@
+package uk.gov.hmcts.reform.iacasedocumentsapi.component.testutils;
+
+import java.util.Comparator;
+
+import org.junit.jupiter.api.ClassDescriptor;
+import org.junit.jupiter.api.ClassOrderer;
+import org.junit.jupiter.api.ClassOrdererContext;
+
+public class TestClassOrderer implements ClassOrderer {
+    @Override
+    public void orderClasses(ClassOrdererContext context) {
+        context.getClassDescriptors().sort(Comparator.comparing((ClassDescriptor d) -> d.getTestClass().getSimpleName()));
+        System.out.println("Applied custom test class execution order. Execution order follows:");
+        context.getClassDescriptors().forEach(descriptor -> System.out.println("    " + descriptor.getTestClass().getName()));
+    }
+}
+
