+++ src/integrationTest/resources/junit-platform.properties
@@ -1,0 +1,1 @@
+junit.jupiter.testclass.order.default=uk.gov.hmcts.reform.iacasedocumentsapi.component.testutils.TestClassOrderer
