+++ src/main/java/uk/gov/hmcts/reform/iacasedocumentsapi/infrastructure/clients/DmDocumentManagementUploader.java
@@ -1,0 +1,91 @@
+package uk.gov.hmcts.reform.iacasedocumentsapi.infrastructure.clients;
+
+import com.google.common.io.ByteStreams;
+import java.io.IOException;
+import java.util.Collections;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.core.io.Resource;
+import org.springframework.stereotype.Component;
+import org.springframework.web.multipart.MultipartFile;
+import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
+import uk.gov.hmcts.reform.document.DocumentUploadClientApi;
+import uk.gov.hmcts.reform.document.domain.UploadResponse;
+import uk.gov.hmcts.reform.document.utils.InMemoryMultipartFile;
+import uk.gov.hmcts.reform.iacasedocumentsapi.domain.UserDetailsProvider;
+import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.UserDetails;
+import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.ccd.field.Document;
+
+
+/**
+ * Superseded. Will need to be removed as soon as the "use-ccd-document-am" feature flag is permanently on
+ */
+
+@Component
+@Deprecated
+public class DmDocumentManagementUploader {
+
+    private final DocumentUploadClientApi documentUploadClientApi;
+    private final AuthTokenGenerator serviceAuthorizationTokenGenerator;
+    private final UserDetailsProvider userDetailsProvider;
+
+    public DmDocumentManagementUploader(
+        DocumentUploadClientApi documentUploadClientApi,
+        AuthTokenGenerator serviceAuthorizationTokenGenerator,
+        @Qualifier("requestUser") UserDetailsProvider userDetailsProvider
+    ) {
+        this.documentUploadClientApi = documentUploadClientApi;
+        this.serviceAuthorizationTokenGenerator = serviceAuthorizationTokenGenerator;
+        this.userDetailsProvider = userDetailsProvider;
+    }
+
+    public Document upload(
+        Resource resource,
+        String contentType
+    ) {
+        final String serviceAuthorizationToken = serviceAuthorizationTokenGenerator.generate();
+        final UserDetails userDetails = userDetailsProvider.getUserDetails();
+        final String accessToken = userDetails.getAccessToken();
+        final String userId = userDetails.getId();
+
+        try {
+
+            MultipartFile file = new InMemoryMultipartFile(
+                resource.getFilename(),
+                resource.getFilename(),
+                contentType,
+                ByteStreams.toByteArray(resource.getInputStream())
+            );
+
+            UploadResponse uploadResponse =
+                documentUploadClientApi
+                    .upload(
+                        accessToken,
+                        serviceAuthorizationToken,
+                        userId,
+                        Collections.singletonList(file)
+                    );
+
+            uk.gov.hmcts.reform.document.domain.Document uploadedDocument =
+                uploadResponse
+                    .getEmbedded()
+                    .getDocuments()
+                    .get(0);
+
+            return new Document(
+                uploadedDocument
+                    .links
+                    .self
+                    .href,
+                uploadedDocument
+                    .links
+                    .binary
+                    .href,
+                uploadedDocument
+                    .originalDocumentName
+            );
+
+        } catch (IOException e) {
+            throw new IllegalStateException(e);
+        }
+    }
+}
