+++ src/test/java/uk/gov/hmcts/reform/iacasedocumentsapi/infrastructure/clients/DocumentManagementUploaderTest.java
@@ -1,121 +1,62 @@
 package uk.gov.hmcts.reform.iacasedocumentsapi.infrastructure.clients;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.Mockito.*;
-
-import java.io.ByteArrayInputStream;
 import java.io.IOException;
-import java.io.InputStream;
-import java.util.List;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Captor;
+import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.core.io.Resource;
-import org.springframework.web.multipart.MultipartFile;
-import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.document.DocumentUploadClientApi;
-import uk.gov.hmcts.reform.document.domain.UploadResponse;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.UserDetailsProvider;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.UserDetails;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.ccd.field.Document;
+import uk.gov.hmcts.reform.iacasedocumentsapi.domain.service.FeatureToggler;
+
+import static org.mockito.ArgumentMatchers.anyBoolean;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.eq;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
 @SuppressWarnings("unchecked")
-public class DocumentManagementUploaderTest {
+class DocumentManagementUploaderTest {
 
-    @Mock private DocumentUploadClientApi documentUploadClientApi;
-    @Mock private AuthTokenGenerator serviceAuthorizationTokenGenerator;
-    @Mock private UserDetailsProvider userDetailsProvider;
+    @Mock
+    private FeatureToggler featureToggler;
+    @Mock
+    private DmDocumentManagementUploader dmDocumentManagementUploader;
+    @Mock
+    private CdamDocumentManagementUploader cdamDocumentManagementUploader;
 
-    private String serviceAuthorizationToken = "SERVICE_TOKEN";
-    private String accessToken = "ACCESS_TOKEN";
-    private String userId = "123";
-    @Mock private UploadResponse uploadResponse;
-    @Mock private UploadResponse.Embedded uploadResponseEmbedded;
-    @Mock private List<uk.gov.hmcts.reform.document.domain.Document> uploadedDocuments;
-    private uk.gov.hmcts.reform.document.domain.Document uploadedDocument
-        = new uk.gov.hmcts.reform.document.domain.Document();
-
-    private String contentType = "application/pdf";
-    private String fileName = "some-file.pdf";
-
-    private byte[] documentData = "pdf-data".getBytes();
-    @Mock private Resource resource;
-    private InputStream resourceInputStream = new ByteArrayInputStream(documentData);
-    private String expectedDocumentUrl = "document-self-href";
-    private String expectedBinaryUrl = "document-binary-href";
-
-    @Mock private UserDetails userDetails;
-
-    @Captor private ArgumentCaptor<List<MultipartFile>> multipartFilesCaptor;
-
+    @InjectMocks
     private DocumentManagementUploader documentManagementUploader;
 
     @BeforeEach
     public void setUp() {
 
-        documentManagementUploader =
-            new DocumentManagementUploader(
-                documentUploadClientApi,
-                serviceAuthorizationTokenGenerator,
-                userDetailsProvider
-            );
-
-        uploadedDocument.originalDocumentName = fileName;
-        uploadedDocument.links = new uk.gov.hmcts.reform.document.domain.Document.Links();
-        uploadedDocument.links.self = new uk.gov.hmcts.reform.document.domain.Document.Link();
-        uploadedDocument.links.self.href = expectedDocumentUrl;
-        uploadedDocument.links.binary = new uk.gov.hmcts.reform.document.domain.Document.Link();
-        uploadedDocument.links.binary.href = expectedBinaryUrl;
     }
 
     @Test
-    public void should_upload_document_to_document_management_and_return_links() throws IOException {
+    void should_use_cdam_when_feature_flag_true() throws IOException {
+        // Given
+        given(featureToggler.getValue(eq("use-ccd-document-am"), anyBoolean())).willReturn(true);
 
-        when(serviceAuthorizationTokenGenerator.generate()).thenReturn(serviceAuthorizationToken);
-        when(userDetails.getAccessToken()).thenReturn(accessToken);
-        when(userDetails.getId()).thenReturn(userId);
-        when(userDetailsProvider.getUserDetails()).thenReturn(userDetails);
+        // When
+        documentManagementUploader.upload(null, null);
 
-        when(resource.getFilename()).thenReturn(fileName);
-        when(resource.getInputStream()).thenReturn(resourceInputStream);
-
-        when(uploadResponse.getEmbedded()).thenReturn(uploadResponseEmbedded);
-        when(uploadResponseEmbedded.getDocuments()).thenReturn(uploadedDocuments);
-        when(uploadedDocuments.get(0)).thenReturn(uploadedDocument);
-
-        when(documentUploadClientApi.upload(
-            eq(accessToken),
-            eq(serviceAuthorizationToken),
-            eq(userId),
-            any(List.class)
-        )).thenReturn(uploadResponse);
-
-        final Document actualDocument = documentManagementUploader.upload(
-            resource,
-            contentType
-        );
-
-        assertEquals(fileName, actualDocument.getDocumentFilename());
-        assertEquals(expectedDocumentUrl, actualDocument.getDocumentUrl());
-        assertEquals(expectedBinaryUrl, actualDocument.getDocumentBinaryUrl());
-
-        verify(documentUploadClientApi, times(1)).upload(
-            eq(accessToken),
-            eq(serviceAuthorizationToken),
-            eq(userId),
-            multipartFilesCaptor.capture()
-        );
-
-        List<MultipartFile> actualMultipartFiles = multipartFilesCaptor.getAllValues().get(0);
-
-        assertEquals(1, actualMultipartFiles.size());
-        assertEquals(fileName, actualMultipartFiles.get(0).getName());
-        assertEquals(fileName, actualMultipartFiles.get(0).getOriginalFilename());
-        assertEquals(documentData.length, actualMultipartFiles.get(0).getBytes().length);
+        // Then
+        verify(cdamDocumentManagementUploader, times(1)).upload(null, null);
     }
+
+    @Test
+    void should_use_dm_when_feature_flag_false() throws IOException {
+        // Given
+        given(featureToggler.getValue(eq("use-ccd-document-am"), anyBoolean())).willReturn(false);
+
+        // When
+        documentManagementUploader.upload(null, null);
+
+        // Then
+        verify(dmDocumentManagementUploader, times(1)).upload(null, null);
+    }
+
 }
